# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Система управления мероприятиями: события, площадки, билеты, отчёты, аудит.",
        "title": "Time2Meet API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tickets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Список билетов по покупателю",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Buyer ID (UUID)",
                        "name": "buyer_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.TicketSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tickets/purchase": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Купить билет (транзакция)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) for audit",
                        "name": "X-User-Id",
                        "in": "header"
                    },
                    {
                        "description": "Покупка",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PurchaseTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.TicketIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Получить билет по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TicketSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tickets"
                ],
                "summary": "Удалить билет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tickets/{id}/status": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Обновить статус билета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Статус",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateTicketStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tickets/{id}/validate": {
            "post": {
                "tags": [
                    "tickets"
                ],
                "summary": "Валидация (использование) билета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) for audit",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ticket ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.Code": {
            "type": "string",
            "enum": [
                "not_found",
                "conflict",
                "validation",
                "unauthorized",
                "forbidden",
                "internal",
                "invalid_state",
                "unavailable"
            ],
            "x-enum-varnames": [
                "CodeNotFound",
                "CodeConflict",
                "CodeValidation",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeInternal",
                "CodeInvalidState",
                "CodeUnavailable"
            ]
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.Code"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PurchaseTicketRequest": {
            "type": "object",
            "required": [
                "amount_paid",
                "qr_code",
                "ticket_type_id"
            ],
            "properties": {
                "amount_paid": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "qr_code": {
                    "type": "string"
                },
                "ticket_type_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.TicketIDResponse": {
            "type": "object",
            "properties": {
                "ticket_id": {
                    "type": "string"
                }
            }
        },
        "handler.TicketSwagger": {
            "type": "object",
            "properties": {
                "amountPaid": {
                    "$ref": "#/definitions/valueobject.Money"
                },
                "buyerID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "purchaseDate": {
                    "type": "string"
                },
                "qrcode": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/valueobject.TicketStatus"
                },
                "ticketTypeID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usedAt": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateTicketStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "valueobject.Money": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "valueobject.TicketStatus": {
            "type": "string",
            "enum": [
                "paid",
                "refunded",
                "void",
                "used"
            ],
            "x-enum-varnames": [
                "TicketStatusPaid",
                "TicketStatusRefunded",
                "TicketStatusVoid",
                "TicketStatusUsed"
            ]
        }
    }
}